backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
openai==1.3.0
boto3==1.34.0
azure-identity==1.15.0
azure-mgmt-resource==23.0.1
azure-mgmt-compute==30.0.0
azure-mgmt-storage==21.0.0
requests==2.31.0
pydantic==2.5.0
mcp==0.1.0
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0
